requirements.txt
fastapi
uvicorn
slowapi
python-dotenv
sqlalchemy
bcrypt
python-jose[cryptography]
python-multipart
psutil
watchfiles
aiosqlite
